% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_functions.R
\name{parse.raw.email}
\alias{parse.raw.email}
\title{This functions parses the raw emali to metadata and body}
\usage{
parse.raw.email(raw_email)
}
\arguments{
\item{raw_emal}{string}
}
\value{
list of metadata and body
}
\description{
This functions parses the raw emali to metadata and body
}
